Decouple model from view, add CI and testability infrastructure
Removes the view dependency from the model layer and adds the tooling to
keep the codebase honest. 64 tests passing, clean under strict warnings
and sanitizers.

Model / view separation:
- Player performed console output: it printed its own bet, fold and hand
  text. It now holds state only and returns handToString() for the view
  to render.
- BotPlayer made 22 direct calls into BotThinkingVisualizer, so the model
  could not think without a console attached and tests were forced to
  render simulation panels to stdout. It now reports through a
  BotObserver interface; BotThinkingReporter adapts that to the console
  view, and the controller wires the two together. Passing no observer
  leaves the bot silent, which is what tests want.
- The model layer now has zero references to view/.

Testability:
- The controller read std::cin inline, so the round flow - pot accounting,
  folds, all-ins, uncalled bets - was unreachable from a test. Input now
  comes through a PlayerInput interface with console and scripted
  implementations, and playSingleRound exposes a single hand.
- BotPlayer gained a seeded constructor so randomised branches reproduce.
- The HardPlus simulation count and the controller's cosmetic thinking
  delay are both configurable. The 2s delay dwarfed the ~10ms decision it
  was covering and dominated test runtime; the suite went from 66s to 12s.

Infrastructure:
- GitHub Actions running the build, the test suite, a strict warning pass
  and an ASan/UBSan pass. 64 tests existed with nothing running them.
- .gitignore added and the 515KB poker binary untracked.
- `make strict` (-Wshadow -Wconversion -Wsign-conversion and friends) and
  `make test-asan` targets, replacing hand-written sanitizer invocations.
  Fixed the four narrowing and sign-conversion warnings this surfaced.

Documentation:
- The docs described 200 simulations across 4 threads with a +/-7% margin.
  All three claims were stale: it now runs 2000 trials single-threaded at
  +/-1.4%. Corrected in BOT_DECISION_FLOW, MONTE_CARLO_ANALYSIS and
  SYSTEM_ARCHITECTURE.
- Documented the dominant limitation with measurements: the simulator
  assumes a uniformly random opponent holding, overstating equity by ~13
  points against a top-25% range and flipping roughly one decision in six.
- readme.md claimed "clean separation of model, controller, view" which
  was not true when written. It is now, and the known limitations (no
  blinds, no preflop round, no raises) are stated.

+++ model/advanced_hand_evaluator.cpp
@@ -219,7 +219,7 @@ HandValue AdvancedHandEvaluator::evaluate(const std::vector<Card> &cards)
     if (pairs.size() == 1)
     {
         std::vector<int> kickers = {pairs[0]};
-        for (int i = 0; i < 3 && i < (int)singles.size(); ++i)
+        for (size_t i = 0; i < 3 && i < singles.size(); ++i)
             kickers.push_back(singles[i]);
         return HandValue{HandRank::OnePair, kickers};
     }
